:recycle: refactor: Member DTO, VO 리팩토링
