+++ src/main/java/sandbox/apricot/auth/dto/MemberPrincipalDetails.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-import sandbox.apricot.member.entity.Member;
+import sandbox.apricot.member.vo.Member;
 
 @Getter
 @RequiredArgsConstructor
@@ -33,23 +33,4 @@ public class MemberPrincipalDetails implements UserDetails {
         return member.getEmail();
     }
 
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
 }
