+++ src/main/java/sandbox/apricot/member/dto/request/MemberRegisterBasic.java
@@ -1,22 +1,26 @@
 package sandbox.apricot.member.dto.request;
 
 import jakarta.validation.constraints.Email;
-import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import sandbox.apricot.member.util.annotation.NickName;
 import sandbox.apricot.member.util.annotation.Password;
 
 @Getter
-@AllArgsConstructor
+@Builder
+@RequiredArgsConstructor
 public class MemberRegisterBasic {
 
     @Email(message = ">>> [ ❌ 유효하지 않은 이메일 형식입니다. ]")
-    private final String email; // 로그인 아이디
-
+    private final String email;
     @NickName
-    private final String nickName; // 닉네임
-
+    private final String nickName;
     @Password
-    private final String password; // 로그인 비밀번호
+    private final String password;
+
+    public MemberRegisterBasic toService() {
+        return this;
+    }
 
 }
