+++ src/main/java/sandbox/apricot/member/dto/request/MemberRegisterDetail.java
@@ -1,12 +1,14 @@
 package sandbox.apricot.member.dto.request;
 
+import jakarta.validation.constraints.NotNull;
 import java.util.List;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.ToString;
+import lombok.RequiredArgsConstructor;
 
 @Getter
-@ToString
+@Builder
+@RequiredArgsConstructor
 public class MemberRegisterDetail {
 
     private final String ageRange;
@@ -14,17 +16,12 @@ public class MemberRegisterDetail {
     private final String career;
     private final String marriedStatus;
     private final String numChild;
+
+    @NotNull(message = ">>> [ ❌ 관심사는 최소 1개를 선택해야 합니다. ]")
     private final List<String> interests;
 
-    @Builder
-    public MemberRegisterDetail(String ageRange, String gender, String career,
-            String marriedStatus, String numChild, List<String> interests) {
-        this.ageRange = ageRange;
-        this.gender = gender;
-        this.career = career;
-        this.marriedStatus = marriedStatus;
-        this.numChild = numChild;
-        this.interests = interests;
+    public MemberRegisterDetail toService() {
+        return this;
     }
 
 }
