+++ src/main/java/sandbox/apricot/member/vo/MemberRole.java
@@ -1,4 +1,4 @@
-package sandbox.apricot.member.entity;
+package sandbox.apricot.member.vo;
 
 import lombok.Getter;
